Remove the ethernet<->ap bridge from the mesh.
If you want to put single eth ports into the mesh (e.g. to connect freifunk
routers), create interfaces for them and put these in the mesh. For the
ethernet<->ap bridge to be able to use the batman routing, they have to be put
into a common bridge, cf
http://www.open-mesh.org/wiki/batman-adv/Quick-start-guide